refactor: split format and unpack modules
